refactor(forms): Remove duplicate file
