+++ src/EditModel/option-set/OptionDialogForOptions/OptionDialogForOptions.component.js
@@ -5,7 +5,7 @@ import withStateFrom from 'd2-ui/lib/component-helpers/withStateFrom';
 import { getInstance } from 'd2/lib/d2';
 
 import AddOptionDialog from './AddOptionDialog.component';
-import { createFieldConfigForModelTypes } from '../../form-helpers/modelFieldConfigCreator';
+import { createFieldConfigForModelTypes } from '../../formHelpers';
 import { isAttribute, isFieldCode } from '../../form-helpers/fieldChecks';
 import { optionDialogStore } from '../stores';
 import modelToEditStore from '../../modelToEditStore';
